Modularized UserInfo component. Created two new components, practitioner data form and visitor data form. Added docstrings, to assist in understanding functionality.:
